+++ src/lightning/fabric/strategies/strategy.py
@@ -265,6 +265,12 @@ class Strategy(ABC):
         """Returns model state."""
         return module.state_dict()
 
+    def load_module_state_dict(
+        self, module: Module, state_dict: Dict[str, Union[Any, Tensor]], strict: bool = True
+    ) -> None:
+        """Loads the given state into the model."""
+        module.load_state_dict(state_dict, strict=strict)
+
     def get_optimizer_state(self, optimizer: Optimizer) -> Dict[str, Tensor]:
         """Returns state of an optimizer.
 
@@ -282,15 +288,20 @@ class Strategy(ABC):
     def load_checkpoint(
         self,
         path: _PATH,
-        state: Optional[Dict[str, Union[Module, Optimizer, Any]]] = None,
+        state: Optional[Union[Module, Optimizer, Dict[str, Union[Module, Optimizer, Any]]]] = None,
         strict: bool = True,
     ) -> Dict[str, Any]:
         """Load the contents from a checkpoint and restore the state of the given objects.
 
         Args:
             path: A path to where the file is located
-            state: A dictionary of objects whose state will be restored in-place from the checkpoint path.
-                If no state is given, then the checkpoint will be returned in full.
+            state: Can be one of:
+
+                - A dictionary of objects whose state will be restored in-place from the checkpoint path.
+                - ``None`` or the empty dict: The loaded checkpoint will be returned in full.
+                - A :class:`~torch.nn.Module` instance, if the checkpoint file contains a raw module state dict.
+                - A :class:`~torch.optim.Optimizer` instance, if the checkpoint file contains a raw optimizer state.
+
             strict: Whether to enforce that the keys in `state` match the keys in the checkpoint.
 
         Returns:
@@ -302,13 +313,21 @@ class Strategy(ABC):
         if not state:
             return checkpoint
 
+        if isinstance(state, Module):
+            self.load_module_state_dict(module=state, state_dict=checkpoint, strict=strict)
+            return {}
+
+        if isinstance(state, Optimizer):
+            state.load_state_dict(checkpoint)
+            return {}
+
         _validate_keys_for_strict_loading(state.keys(), checkpoint.keys(), strict=strict)
         for name, obj in state.copy().items():
             if name not in checkpoint:
                 continue
             if isinstance(obj, _Stateful):
                 if isinstance(obj, Module):
-                    obj.load_state_dict(checkpoint.pop(name), strict=strict)
+                    self.load_module_state_dict(module=obj, state_dict=checkpoint.pop(name), strict=strict)
                 else:
                     obj.load_state_dict(checkpoint.pop(name))
             else: